Fix a few -Wconversion warnings (#430)
Wanted to play around with it, and we have a lot of issues. Some of them
are pretty bad too. uint -> int -> uint type stuff. Crazy!

Obviously it is hard to throw -Wconversion on all at once, so here is an
attempt at fixing a few issues that were pretty easy to reason about.

+++ lib/binding/kvdb_interface.c
@@ -75,7 +75,7 @@ hse_lowmem_adjust(unsigned long *memgb)
     hse_meminfo(NULL, &mavail, 30);
 
     if (mavail <= HSE_LOWMEM_THRESHOLD_GB_DFLT) {
-        uint32_t scale = ikvdb_lowmem_scale(mavail);
+        unsigned long scale = ikvdb_lowmem_scale(mavail);
 
         /* Scale various caches based on the available memory */
         if (hse_gparams.gp_c0kvs_ccache_sz_max == gpdef.gp_c0kvs_ccache_sz_max)
